rename DiameterModified -> HistogramAlg
