+++ HistogramAlg.py
